refactor: unify actions remote data source
