+++ composeApp/src/androidMain/kotlin/com/juanpablo0612/sigat/PlatformDi.kt
@@ -1,9 +1,5 @@
 package com.juanpablo0612.sigat
 
-import com.juanpablo0612.sigat.data.actions.remote.ActionsRemoteDataSource
-import com.juanpablo0612.sigat.data.actions.remote.AndroidActionsRemoteDataSource
 import org.koin.dsl.module
 
-actual val platformModule = module {
-    single<ActionsRemoteDataSource> { AndroidActionsRemoteDataSource(get(), get()) }
-}
+actual val platformModule = module { }
