+++ composeApp/src/commonMain/kotlin/com/juanpablo0612/sigat/data/actions/remote/ActionsRemoteDataSource.kt
@@ -3,6 +3,10 @@ package com.juanpablo0612.sigat.data.actions.remote
 import com.juanpablo0612.sigat.data.actions.model.ActionModel
 import com.juanpablo0612.sigat.data.exceptions.handleExceptions
 import dev.gitlive.firebase.firestore.FirebaseFirestore
+import dev.gitlive.firebase.storage.Data
+import dev.gitlive.firebase.storage.FirebaseStorage
+import io.github.vinceglb.filekit.FileKit
+import io.github.vinceglb.filekit.ImageFormat
 import io.github.vinceglb.filekit.PlatformFile
 import io.github.vinceglb.filekit.readBytes
 import kotlinx.coroutines.flow.Flow
@@ -16,7 +20,8 @@ interface ActionsRemoteDataSource {
     fun getActions(creatorUid: String): Flow<List<ActionModel>>
 }
 
-abstract class BaseActionsRemoteDataSource(
+class ActionsRemoteDataSourceImpl(
+    private val storage: FirebaseStorage,
     firestore: FirebaseFirestore,
 ) : ActionsRemoteDataSource {
     private val actionsCollection = firestore.collection("actions")
@@ -44,9 +49,30 @@ abstract class BaseActionsRemoteDataSource(
         }
     }
 
+    override suspend fun uploadImage(
+        actionId: String,
+        imageName: String,
+        imageByteArray: ByteArray
+    ): String {
+        val compressedBytes = compressImage(imageByteArray)
+        val imageRef = storage.reference.child("actions/$actionId/$imageName")
+        imageRef.putData(Data(compressedBytes))
+        return imageRef.getDownloadUrl()
+    }
+
     override fun getActions(creatorUid: String): Flow<List<ActionModel>> {
         return actionsCollection.where {
             "creatorUid" equalTo creatorUid
         }.snapshots.map { it.documents.map { document -> document.data(ActionModel.serializer()) } }
+    }
+
+    private suspend fun compressImage(imageBytes: ByteArray): ByteArray {
+        return FileKit.compressImage(
+            bytes = imageBytes,
+            quality = 80,
+            maxWidth = 1024,
+            maxHeight = 1024,
+            imageFormat = ImageFormat.JPEG
+        )
     }
 }
