--- composeApp/src/desktopMain/kotlin/com/juanpablo0612/sigat/data/actions/remote/DesktopActionsRemoteDataSource.kt
@@ -1,14 +1,0 @@
-package com.juanpablo0612.sigat.data.actions.remote
-
-import dev.gitlive.firebase.firestore.FirebaseFirestore
-
-class DesktopActionsRemoteDataSource(firestore: FirebaseFirestore) :
-    BaseActionsRemoteDataSource(firestore) {
-    override suspend fun uploadImage(
-        actionId: String,
-        imageName: String,
-        imageByteArray: ByteArray
-    ): String {
-        TODO("Not yet implemented")
-    }
-}
